Add migration for system user in AspNetUsers table
This commit introduces a new migration named `AddSystemUser` to the `AutenticazioneSvc` project. It defines the model for the `AspNetUsers` table, including properties such as `Id`, `FirstName`, `LastName`, `UserName`, and `Email`, along with various user account management flags.

The migration logic in `20250216163032_AddSystemUser.cs` executes a SQL script to insert a new user with the username 'System' and email 'system@example.com' during the `Up` method. The `Down` method allows for the removal of this user if the migration is rolled back. Additionally, the SQL script `01_create_system_user.sql` contains the command to perform the insertion.